refactor: recreate droplet on workspace stop/start
